+++ kern/h/kpi/types.h
@@ -49,8 +49,8 @@ typedef size_t          ino_t;
 #endif
 
 
-typedef unsigned int    fs_ftype_t;
-typedef unsigned int    fs_perms_t;
+typedef unsigned short  fs_ftype_t;
+typedef unsigned short  fs_perms_t;
 typedef int             nlink_t;
